Refactoring ImageFileDialog using an abstract class for data dialog
